+++ src/Oro/Bundle/CommerceBundle/Form/Type/ScorecardSelectType.php
@@ -2,7 +2,7 @@
 
 namespace Oro\Bundle\CommerceBundle\Form\Type;
 
-use Oro\Bundle\CommerceBundle\ContentWidget\Provider\CustomerDashboardDatagridsProviderInterface;
+use Oro\Bundle\CommerceBundle\ContentWidget\Provider\ScorecardsRegistryInterface;
 use Oro\Bundle\FormBundle\Form\Type\Select2ChoiceType;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\FormInterface;
@@ -10,12 +10,12 @@ use Symfony\Component\Form\FormView;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
 /**
- * Covers logic of selecting customer dashboard datagrid for content widget.
+ * Covers logic of selecting scorecards for content widget.
  */
-class CustomerDashboardDatagridSelectType extends AbstractType
+class ScorecardSelectType extends AbstractType
 {
     public function __construct(
-        private CustomerDashboardDatagridsProviderInterface $datagridContentWidgetTypeProvider
+        private readonly ScorecardsRegistryInterface $scorecardsRegistry
     ) {
     }
 
@@ -24,7 +24,7 @@ class CustomerDashboardDatagridSelectType extends AbstractType
     {
         if (empty($options['choices'])) {
             $options['configs']['placeholder'] =
-                'oro.commerce.content_widget_type.customer_dashboard_datagrid.form.no_available_datagrids';
+                'oro.commerce.content_widget_type.scorecard.form.no_available_scorecards';
         }
 
         $view->vars = \array_replace_recursive($view->vars, ['configs' => $options['configs']]);
@@ -35,8 +35,8 @@ class CustomerDashboardDatagridSelectType extends AbstractType
     {
         $resolver->setDefaults(
             [
-                'choices' => $this->datagridContentWidgetTypeProvider->getDatagrids(),
-                'placeholder' => 'oro.commerce.content_widget_type.customer_dashboard_datagrid.form.choose_datagrid',
+                'choices' => $this->getScorecardChoices(),
+                'placeholder' => 'oro.commerce.content_widget_type.scorecard.form.choose_scorecard',
             ]
         );
     }
@@ -50,6 +50,16 @@ class CustomerDashboardDatagridSelectType extends AbstractType
     #[\Override]
     public function getBlockPrefix(): string
     {
-        return 'oro_commerce_datagrid_content_widget_type_select';
+        return 'oro_commerce_scorecard_content_widget_type_select';
+    }
+
+    private function getScorecardChoices(): array
+    {
+        $choices = [];
+        foreach ($this->scorecardsRegistry->getProviders() as $provider) {
+            $choices[$provider->getLabel()] = $provider->getName();
+        }
+
+        return $choices;
     }
 }