+++ src/Oro/Bundle/CommerceBundle/Tests/Unit/Layout/DataProvider/PurchaseVolumeChartDataProviderTest.php
@@ -7,11 +7,10 @@ use Oro\Bundle\ChartBundle\Model\ChartView;
 use Oro\Bundle\ChartBundle\Model\ChartViewBuilder;
 use Oro\Bundle\ChartBundle\Model\ConfigProvider;
 use Oro\Bundle\CommerceBundle\Layout\DataProvider\PurchaseVolumeChartDataProvider;
-use Oro\Bundle\CurrencyBundle\Provider\DefaultCurrencyProviderInterface;
 use Oro\Bundle\CustomerBundle\Entity\Customer;
 use Oro\Bundle\CustomerBundle\Entity\CustomerUser;
 use Oro\Bundle\LocaleBundle\Model\LocaleSettings;
-use Oro\Bundle\SecurityBundle\Authentication\TokenAccessorInterface;
+use Oro\Bundle\PricingBundle\Manager\UserCurrencyManager;
 use Oro\Bundle\UserBundle\Entity\User;
 use Oro\Bundle\WebsiteBundle\Entity\Website;
 use Oro\Bundle\WebsiteBundle\Manager\WebsiteManager;
@@ -23,8 +22,7 @@ final class PurchaseVolumeChartDataProviderTest extends TestCase
 {
     private ManagerRegistry&MockObject $registry;
     private WebsiteManager&MockObject $websiteManager;
-    private DefaultCurrencyProviderInterface&MockObject $defaultCurrencyProvider;
-    private TokenAccessorInterface&MockObject $tokenAccessor;
+    private UserCurrencyManager&MockObject $userCurrencyManager;
     private LocaleSettings&MockObject $localeSettings;
     private ConfigProvider&MockObject $configProvider;
     private ChartViewBuilder&MockObject $chartViewBuilder;
@@ -37,8 +35,7 @@ final class PurchaseVolumeChartDataProviderTest extends TestCase
     {
         $this->registry = $this->createMock(ManagerRegistry::class);
         $this->websiteManager = $this->createMock(WebsiteManager::class);
-        $this->defaultCurrencyProvider = $this->createMock(DefaultCurrencyProviderInterface::class);
-        $this->tokenAccessor = $this->createMock(TokenAccessorInterface::class);
+        $this->userCurrencyManager = $this->createMock(UserCurrencyManager::class);
         $this->localeSettings = $this->createMock(LocaleSettings::class);
         $this->configProvider = $this->createMock(ConfigProvider::class);
         $this->chartViewBuilder = $this->createMock(ChartViewBuilder::class);
@@ -61,34 +58,16 @@ final class PurchaseVolumeChartDataProviderTest extends TestCase
         $this->provider = new PurchaseVolumeChartDataProvider(
             $this->registry,
             $this->websiteManager,
-            $this->defaultCurrencyProvider,
-            $this->tokenAccessor,
+            $this->userCurrencyManager,
             $this->localeSettings,
             $this->configProvider,
             $this->chartViewBuilder,
-            ['order_internal_status.cancelled']
+            ['cancelled']
         );
     }
 
     public function testGetPurchaseVolumeChartViewNoWebsite(): void
     {
-        self::assertSame($this->chartView, $this->provider->getPurchaseVolumeChartView());
-    }
-
-    public function testGetPurchaseVolumeChartViewNoCustomer(): void
-    {
-        $this->configProvider->expects(self::once())
-            ->method('getChartConfig')
-            ->with('purchase_volume_chart')
-            ->willReturn([]);
-
-        $website = new Website();
-        ReflectionUtil::setId($website, 1);
-
-        $this->websiteManager->expects(self::once())
-            ->method('getCurrentWebsite')
-            ->willReturn($website);
-
         self::assertSame($this->chartView, $this->provider->getPurchaseVolumeChartView());
     }
 
@@ -108,10 +87,6 @@ final class PurchaseVolumeChartDataProviderTest extends TestCase
 
         $user = new User();
         ReflectionUtil::setId($user, 1);
-
-        $this->tokenAccessor->expects(self::once())
-            ->method('getUser')
-            ->willReturn($user);
 
         self::assertSame($this->chartView, $this->provider->getPurchaseVolumeChartView());
     }
@@ -134,24 +109,20 @@ final class PurchaseVolumeChartDataProviderTest extends TestCase
         ReflectionUtil::setId($customer, 1);
         $customerUser = (new CustomerUser())->setCustomer($customer);
 
-        $this->tokenAccessor->expects(self::once())
-            ->method('getUser')
-            ->willReturn($customerUser);
-
         self::assertSame($this->chartView, $this->provider->getPurchaseVolumeChartView());
     }
 
     public function testSetInternalStatuses(): void
     {
         self::assertSame(
-            ['order_internal_status.cancelled'],
+            ['cancelled'],
             ReflectionUtil::getPropertyValue($this->provider, 'internalStatuses')
         );
 
-        $this->provider->setInternalStatuses(['order_internal_status.open']);
+        $this->provider->setInternalStatuses(['open']);
 
         self::assertSame(
-            ['order_internal_status.open'],
+            ['open'],
             ReflectionUtil::getPropertyValue($this->provider, 'internalStatuses')
         );
     }